Move several functions to child components
