+++ src/views/MainView.vue
@@ -2,38 +2,16 @@
   <div class="main-view">
     <div class="row">
       <div class="col-5">
-        <div class="styled-input">
-          <input
-            class="add-to-do"
-            type="text"
-            placeholder="Add a New Mission..."
-          />
-          <span class="styled-input-append"
-            ><i class="material-icons">add</i></span
-          >
-        </div>
+        <styled-input />
         <template v-if="miniToDoList.length > 0">
           <app-pomodoro-timer />
           <div class="mini-to-do-list">
             <template v-if="miniToDoList.length > 0">
-              <template v-for="todo in miniToDoList">
-                <div class="to-do" :key="todo.title">
-                  <div class="to-do-radio" @click="toggleComplete">
-                    <i class="material-icons unchecked">
-                      radio_button_unchecked
-                    </i>
-                    <i class="material-icons checked">
-                      radio_button_checked
-                    </i>
-                  </div>
-                  <div class="to-do-container">
-                    <div class="to-do-title">{{ todo.title }}</div>
-                    <div class="to-do-icon-start">
-                      <i class="material-icons">play_circle_outline</i>
-                    </div>
-                  </div>
-                </div>
-              </template>
+              <to-do
+                v-for="todo in miniToDoList"
+                :key="todo.title"
+                :title="todo.title"
+              />
             </template>
             <template v-else>
               <div>
@@ -75,12 +53,16 @@
 <script>
 import AppPomodoroTimer from "@/components/AppPomodoroTimer.vue";
 import AppPomodoroPie from "@/components/AppPomodoroPie.vue";
-// import { mapActions } from "vuex";
+import StyledInput from "@/components/StyledInput.vue";
+import ToDo from "@/components/ToDo.vue";
+
 export default {
   name: "MainView",
   components: {
     AppPomodoroTimer,
-    AppPomodoroPie
+    AppPomodoroPie,
+    StyledInput,
+    ToDo
   },
   data: function() {
     return {};
@@ -96,101 +78,12 @@ export default {
       return list;
     }
   },
-  methods: {
-    toggleComplete: function(event) {
-      const el = event.currentTarget;
-      el.classList.contains("done")
-        ? el.classList.remove("done")
-        : el.classList.add("done");
-    }
-  }
+  methods: {}
 };
 </script>
 
 <style lang="scss">
-@import "@/assets/styles/variables.scss";
-@mixin ellipsis {
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-.to-do {
-  display: flex;
-  padding: 10px 0;
-  border-bottom: 1px solid rgba($dark-blue, 0.2);
-}
-.to-do-radio {
-  color: $dark-blue;
-  margin-right: 4px;
-  cursor: pointer;
-  .checked {
-    display: none;
-  }
-  &.done,
-  &:hover {
-    .checked {
-      display: inline-block;
-    }
-    .unchecked {
-      display: none;
-    }
-  }
-}
-.to-do-container {
-  display: flex;
-  flex-grow: 1;
-  min-width: 0;
-}
-.to-do-title {
-  color: $dark-blue;
-  font-weight: 700;
-  text-transform: uppercase;
-  line-height: 1.5rem;
-  flex: 1;
-  @include ellipsis;
-}
-.to-do-icon-start {
-  color: $dark-blue;
-}
-
-.styled-input {
-  position: relative;
-  input[type="text"] {
-    width: 100%;
-    padding: 1.125em 1em;
-    border: 0;
-    &::placeholder {
-      color: $pink;
-      font-weight: 700;
-      font-style: italic;
-      text-transform: uppercase;
-    }
-  }
-  &-append {
-    position: absolute;
-    right: 1rem;
-    top: 50%;
-    transform: translateY(-50%);
-    background-color: $white;
-    // font-size: 0;
-    color: $pink;
-    &:hover {
-      cursor: pointer;
-      // color: $dark-blue;
-    }
-  }
-}
-.break {
-  .styled-input {
-    input[type="text"]::placeholder {
-      color: $blue;
-    }
-  }
-  .styled-input-append {
-    color: $blue;
-  }
-}
+// @import "@/assets/styles/variables.scss";
 
 .nav {
   display: flex;
